# shared/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SeoProjectDetailOut": {
        "properties": {
          "audits": {
            "items": {
              "$ref": "#/components/schemas/SiteAuditSummaryOut"
            },
            "title": "Audits",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "domain": {
            "title": "Domain",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "latest_audit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SiteAuditSummaryOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "domain",
          "created_at",
          "updated_at",
          "latest_audit",
          "audits"
        ],
        "title": "SeoProjectDetailOut",
        "type": "object"
      },
...
      "SiteAuditSummaryOut": {
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_step": {
            "anyOf": [
              {
                "enum": [
                  "discovery",
                  "crawl",
                  "analysis",
                  "finalize"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Step"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "health_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Health Score"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "js_rendering": {
            "title": "Js Rendering",
            "type": "boolean"
          },
          "page_limit": {
            "title": "Page Limit",
            "type": "integer"
          },
          "pages_crawled": {
            "title": "Pages Crawled",
            "type": "integer"
          },
          "pages_discovered": {
            "title": "Pages Discovered",
            "type": "integer"
          },
          "profile_id": {
            "enum": [
              "site_audit_mobile",
              "site_audit_desktop"
            ],
            "title": "Profile Id",
            "type": "string"
          },
          "progress": {
            "title": "Progress",
            "type": "integer"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "done",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "total_errors": {
            "title": "Total Errors",
            "type": "integer"
          },
          "total_notices": {
            "title": "Total Notices",
            "type": "integer"
          },
          "total_warnings": {
            "title": "Total Warnings",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "status",
          "progress",
          "current_step",
          "error",
          "page_limit",
          "profile_id",
          "js_rendering",
          "pages_discovered",
          "pages_crawled",
          "total_errors",
          "total_warnings",
          "total_notices",
          "health_score",
          "created_at",
          "updated_at",
          "started_at",
          "completed_at"
        ],
        "title": "SiteAuditSummaryOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Yanki API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/seo-projects/{project_id}": {
      "get": {
        "operationId": "read_seo_project_api_v1_seo_projects__project_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Org-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeoProjectDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Seo Project",
        "tags": [
          "seo-projects"
        ]
      }
    },
...
  }
}